:pencil2: fix
modified:   project/server/functions.py; modified:   project/server/main/views.py; modified:   project/server/tasks.py; modified:   ../backend/modules/api.py; modified:   ../backend/modules/run.py; deleted:    ../backend/utilities/utilities.py; modified:   ../frontend/src/app/app-routing.module.ts; modified:   ../frontend/src/app/app.module.ts; modified:   ../frontend/src/app/db.service.ts; new file:   ../frontend/src/app/job/job-history/job-history.component.css; new file:   ../frontend/src/app/job/job-history/job-history.component.html; new file:   ../frontend/src/app/job/job-history/job-history.component.spec.ts; new file:   ../frontend/src/app/job/job-history/job-history.component.ts; new file:   ../frontend/src/app/job/job-result/job-result.component.html; new file:   ../frontend/src/app/job/job-result/job-result.component.scss; new file:   ../frontend/src/app/job/job-result/job-result.component.spec.ts; new file:   ../frontend/src/app/job/job-result/job-result.component.ts; modified:   ../frontend/src/app/job/play-job/play-job.component.html; modified:   ../frontend/src/app/job/play-job/play-job.component.ts; modified:   ../frontend/src/app/job/run-job/run-job.component.ts

+++ backend/modules/api.py
@@ -1,11 +1,15 @@
 import json
+from bson import json_util, ObjectId
 from flask import request, jsonify
 from bson.objectid import ObjectId
 from flask_cors import CORS
-
-from utilities.utilities import *
+from pymongo import MongoClient
+import yaml
 # Import Libraries 
 from app import app
+def parse_json(data):
+    return json.loads(json_util.dumps(data))
+db = MongoClient(yaml.load(open('database.yaml'),Loader=yaml.FullLoader)['uri'])['jobilee']
 
 @app.route('/<collection>', methods=['POST', 'GET'])
 def data(collection):
@@ -23,29 +27,34 @@ def data(collection):
         return jsonify(parse_json(dataJson))
 
 @app.route('/<collection>/<string:id>', methods=['GET', 'DELETE', 'PUT'])
-def onedata(collection,id):
+@app.route('/<collection>/<string:key>/<string:value>', methods=['GET'])
+def onedata(collection,id=None,key=None,value=None):
+    if id:
+        # GET all data from database
+        if request.method == 'GET':
+            data = db[collection].find_one({'_id': ObjectId(id)})
+            db_doc = {k: v if k != '_id' else str(v) for k, v in data.items()}
+            db_doc.update({'_id': str(db_doc['_id'])} if '_id' in db_doc else {})    
+            return jsonify(parse_json(db_doc))
+            
+        # DELETE a data
+        if request.method == 'DELETE':
+            db[collection].delete_many({'_id': ObjectId(id)})
+            print('\n # Deletion successful # \n')
+            return jsonify({'status': 'Data id: ' + id + ' is deleted!'})
 
-    # GET all data from database
-    if request.method == 'GET':
-        data = db[collection].find_one({'_id': ObjectId(id)})
-        db_doc = {k: v if k != '_id' else str(v) for k, v in data.items()}
-        db_doc.update({'_id': str(db_doc['_id'])} if '_id' in db_doc else {})    
-        return jsonify(parse_json(db_doc))
-        
-    # DELETE a data
-    if request.method == 'DELETE':
-        db[collection].delete_many({'_id': ObjectId(id)})
-        print('\n # Deletion successful # \n')
-        return jsonify({'status': 'Data id: ' + id + ' is deleted!'})
-
-    # UPDATE a data by id
-    if request.method == 'PUT':
-        update_doc = {k: v for k, v in request.json.items() if k != '_id'}
-        db[collection].update_one(
-            {'_id': ObjectId(id)},
-            { "$set": update_doc }
-        )
-        return jsonify({'status': 'Data id: ' + id + ' is updated!'})
+        # UPDATE a data by id
+        if request.method == 'PUT':
+            update_doc = {k: v for k, v in request.json.items() if k != '_id'}
+            db[collection].update_one(
+                {'_id': ObjectId(id)},
+                { "$set": update_doc }
+            )
+            return jsonify({'status': 'Data id: ' + id + ' is updated!'})
+    else:
+        allData = db[collection].find({key:value})
+        dataJson = [{k: (str(v) if k == '_id' else v) for k, v in data.items()} for data in allData]
+        return jsonify(parse_json(dataJson))
 
 if __name__ == '__main__':
     app.debug = True
